Converted coordinates and dimensions to be parsed as floats rather than integers
Affects location, lot, beacon and building models

+++ app/models/beacon.rb
@@ -8,6 +8,6 @@ class Beacon < ApplicationRecord
 
     private
         def parse_coordinates
-            self.coordinates = self.coordinates.map{|k,v| [k,v.to_i]}.to_h
+            self.coordinates = self.coordinates.map{|k,v| [k,v.to_f]}.to_h
         end
 end
